Add and reorganize tests
