+++ example/recursive/template/file.yaml.j2
@@ -1,0 +1,3 @@
+var0: random data
+varA: {{ varA }}
+varB: {{ varB }}
