--- example/symlink/template/file.yaml.j2
@@ -1,1 +1,0 @@
-my_var: {{ my_var }}
