--- example/symlink/template/link_orig.yaml
@@ -1,1 +1,0 @@
-file.yaml
