+++ test/data/input/minimal
@@ -1,0 +1,1 @@
+../../../example/minimal
