+++ test/data/input/recursive
@@ -1,0 +1,1 @@
+../../../example/recursive
