+++ test/data/input/simple
@@ -1,0 +1,1 @@
+../../../example/simple
