+++ test/data/output/recursive/.fmf/version
@@ -1,0 +1,1 @@
+1
