+++ test/data/output/recursive/A_2/B_x/file.yaml
@@ -1,0 +1,3 @@
+var0: random data
+varA: 2
+varB: x
