+++ test/data/output/recursive/A_3/B_x/file.yaml
@@ -1,0 +1,3 @@
+var0: random data
+varA: 3
+varB: x
