+++ test/data/output/simple/.fmf/version
@@ -1,0 +1,1 @@
+1
