+++ test/data/output/simple/main.fmf
