+++ test/data/output/simple/rootA/fileB.yaml
@@ -1,0 +1,1 @@
+../rootB/file.yaml
