+++ test/data/output/simple/rootA/renamed_file.txt
@@ -1,0 +1,1 @@
+This is a file that is simply copied over.
