+++ test/data/output/simple/rootB/file.yaml
@@ -1,0 +1,3 @@
+var0: random data
+var1: 42
+var2: Overwritten
