--- test/unit/data/input/trees/simple
@@ -1,1 +1,0 @@
-../../../../../example/simple
