--- test/unit/data/input/trees/symlink
@@ -1,1 +1,0 @@
-../../../../../example/symlink
