--- test/unit/data/output/trees/simple/main.fmf
@@ -1,8 +1,0 @@
-templates: /template
-
-/rootA:
-  vars:
-    my_var: A
-/rootB:
-  vars:
-    my_var: B
