--- test/unit/data/output/trees/simple/rootA/file.yaml
@@ -1,1 +1,0 @@
-my_var: A
