--- test/unit/data/output/trees/simple/rootB/file.yaml
@@ -1,1 +1,0 @@
-my_var: B
