--- test/unit/data/output/trees/simple/rootB/some_file.yaml
@@ -1,1 +1,0 @@
-type: normal_file
