--- test/unit/data/output/trees/symlink/main.fmf
@@ -1,12 +1,0 @@
-templates: /template
-
-/rootA:
-  links:
-    link_tpl.yaml: file.yaml
-  vars:
-    my_var: A
-/rootB:
-  links:
-    link_to_A.yaml: /rootA/file.yaml
-  vars:
-    my_var: B
