--- test/unit/data/output/trees/symlink/rootA/link_orig.yaml
@@ -1,1 +1,0 @@
-file.yaml
