--- test/unit/data/output/trees/symlink/rootA/link_tpl.yaml
@@ -1,1 +1,0 @@
-file.yaml
