--- test/unit/data/output/trees/symlink/rootB/file.yaml
@@ -1,1 +1,0 @@
-my_var: B
