--- test/unit/data/output/trees/symlink/rootB/link_orig.yaml
@@ -1,1 +1,0 @@
-file.yaml
