--- test/unit/data/output/trees/symlink/rootB/link_to_A.yaml
@@ -1,1 +1,0 @@
-../rootA/file.yaml
